requirements.txt
asgiref==3.11.1
boto3==1.42.56
botocore==1.42.56
certifi==2026.2.25
charset-normalizer==3.4.4
dj-database-url==3.1.2
Django==4.2.28
django-storages==1.14.6
gunicorn==25.1.0
idna==3.11
jmespath==1.1.0
packaging==26.0
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.11.0
django-ratelimit==4.1.0
